rbh-sync: add lustre-mpi script
This patch adds the test for the LUSTRE-MPI backend. The test is the
same as for the LUSTRE backend so it's run twice with a different
environment variable. The environment variable "WITH_MPI" is set in
the meson.build with the parameter env to tell if we want to run the
test with MPI or not.

Change-Id: Icf4969413a235227120a3b797def60664efa0f98
Signed-off-by: Gauthier EVRAERD <gauthier.evraerd.ocre@cea.fr>