retention: use modify time for directory expiration
This patch changes the time used to calculate an expiration for a
directory from the maximum between atime and mtime to simply the mtime.

Also update the test of retention script to take that change into
account.

Finally, add the management of empty directories when checking the tree
of the expired directories.

Change-Id: I2f5ff700e5683b221081ae946896fc11e3ba100c
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>