lustre-mpi: add branches to lustre-mpi backend
This patch adds the branch to the lustre-mpi backend. Creating a
branch follows the same principle as creating the backend. It uses the
"posix_branch_backend" structure but by overloading it.

Also, add a boolean in the "mpi_iterator" structure to indicate if we
are synchronizing a branch or not. It is used to know if we have to
modify the root's name and parent ID to match Robinhood's conventions.

Finally, add a new test for the Lustre and Lustre-mpi backend with
rbh-sync to test if the branch works correctly. It's the same test as
for the POSIX backend.

Change-Id: Iac85fda2d0924695b0baf2bd85ab242ac0b96ac1
Signed-off-by: Gauthier EVRAERD <gauthier.evraerd.ocre@cea.fr>