mongo backend: get address to remote database from an environment variable
This patch adds the ability to the mongo backend to remotely connect
to a mongo database. The database address is specified in the
"ADDR_DB" environment variable. If the variable is not defined, the
mongo backend connects by default to "localhost:27017" otherwise to the
address specified by the variable. The variable must be "IP:port".

Change-Id: Ic638c0cdb46f1222d9866efe40c315c507270e8c
Signed-off-by: Gauthier EVRAERD <gauthier.evraerd.ocre@cea.fr>