mongo backend: get address to remote database from an environment variable
This patch adds the ability to the mongo backend to remotely connect
to a mongo database. The database address is specified in the
"RBH_MONGO_DB_URI" environment variable. If the variable is not
defined, the mongo backend connects by default to
"mongodb://localhost:27017" otherwise to the address specified by the
variable. The variable must be "mongodb://host:port".

Change-Id: Ic638c0cdb46f1222d9866efe40c315c507270e8c
Signed-off-by: Gauthier EVRAERD <gauthier.evraerd.ocre@cea.fr>
Reviewed-on: https://review.gerrithub.io/c/robinhood-suite/robinhood4/+/1184132
Reviewed-by: VALERI Yoann <yoann.valeri@cea.fr>
Tested-by: Guillaume Courrier <guillaume.courrier@cea.fr>
Reviewed-by: Guillaume Courrier <guillaume.courrier@cea.fr>