rbh backends: add rbh_backend_plugin_destroy
This patch adds the function "rbh_backend_plugin_destroy" and
"rbh_*_plugin_destroy" in each backend. The latter is used to release
the memory associated with a plugin.

For the moment, only rbh-sync calls it because of the LUSTRE-MPI
backend which need to call "MPI_Finalize" only at the end of the
program.

Change-Id: I40a76faa4bcb2d07a0f213e9a95077343787f697
Signed-off-by: Gauthier EVRAERD <gauthier.evraerd.ocre@cea.fr>