rbh backends: add rbh_backend_plugin_destroy
This patch add the function "rbh_backend_plugin_destroy" and
"rbh_*_backend_destroy" at each backend. This latter works as a
wrapper to the destroy function of the backend. It's useful to
finalize of extern library or doing work that is not directly related
to the destruction of the backend structure.

For the moment, only rbh-sync call it because of the LUSTRE-MPI
backend which need to call "MPI_Finalize" only at the end of the
program.

Change-Id: I40a76faa4bcb2d07a0f213e9a95077343787f697
Signed-off-by: Gauthier EVRAERD <gauthier.evraerd.ocre@cea.fr>