backends: add safety on the number of available inode pairs
This patch adds a check when adding new pairs to the inode xattrs, so
that we don't go over the size allocated previously. For this we must
add an argument the inode xattrs callback of the Lustre backend and the
entry point of the Lustre enricher.

Since the number of parameters for these functions is high, and multiple
parameters concern the entry itself, we create a new structure
`entry_info` that will hold these values, and will be passed between the
backends.

Change-Id: Ia91f80208f784c3eebdd7851b40d6239917f570d
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>