fsevents: keep mdt index of symlink in fsevents
This patch solves an issue regarding the recording of a symlink's MDT
index by using the MDT provided for a rbh-fsevents invocation.

The issue was that Lustre doesn't allow us to get the MDT index of a
symlink because the underlying ioctl done for a getstripe fails. This
means we cannot get the MDT index when synchronizing with rbh-sync.

To circumvent this issue, we instead rely on rbh-fsevents, and more
specifically the fact that if we detect a symlink's creation, its MDT
index is the index of the MDT being monitored. Therefore, we can
translate the MDT name given to rbh-fsevents to a MDT index, and assign
this value to the symlink's MDT index.

Change-Id: I6a7ff7f82f614ae0ee548ab56e8ee827f1975adc
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>