mpi-iter: move the mpi iterator/backend functions
This patch moves the mpi iterator/backend functions originally placed
in the lustre-mpi backend in a folder containing all the functions
useful for mpi. As it was before, we had to have Lustre installed and
the lustre-mpi backend to have the posix-mpi backend. Also, this adds
better generic management of the mpi backends.

Change-Id: I18ed9353e2344a54f9d210aaa238ad194215d9eb
Signed-off-by: Gauthier EVRAERD <gauthier.evraerd.ocre@cea.fr>