posix-mpi: add the new posix-mpi backend
This patch adds the new posix-mpi backend and his features. It is done
the same way as the lustre-mpi backend and use the functions from
mpi-iter. The only difference is that it has no callback and no
"get_attribute" function.

As the other backends, it is based on the POSIX backend but overloads
it.

Change-Id: I6a7f611eb1b92039179e6e0771723ed1fddd047d
Signed-off-by: Gauthier EVRAERD <gauthier.evraerd.ocre@cea.fr>
Reviewed-on: https://review.gerrithub.io/c/robinhood-suite/robinhood4/+/1193355
Reviewed-by: Guillaume Courrier <guillaume.courrier@cea.fr>
Reviewed-by: VALERI Yoann <yoann.valeri@cea.fr>
Tested-by: VALERI Yoann <yoann.valeri@cea.fr>