posix-mpi: add unit tests for the posix-mpi backend
This patch adds the unit tests for the posix-mpi backend. As for the
lustre-mpi backend, we reuse the same tests as the posix backend
because we want to have the same behavior with mpifileutils. But since
we use mpi, we have several processes, so we need to modify a bit the
test with an empty root. We need to check that only one process flist
is not empty and for the others processes, we need to check that the
iterator returns NULL and set errno to ENODATA.

Change-Id: I8b5b6dc7c91d8beb51b50db7862b75a91f8d0e66
Signed-off-by: Gauthier EVRAERD <gauthier.evraerd.ocre@cea.fr>