lustre: add `dir.lov` to `lustre_backend_get_attribute`
This patch adds a way to retrieve a directory's data striping from
outside of the Lustre backend by using the
`lustre_backend_get_attribute` function.

This function is already used by the Lustre enricher, but this patch
adds a valid attribute name, which is `dir.lov`, and then check the key
of the rbh_value_pair given as parameter.

According to that key, we retrieve a certain value from a directory's
striping, and put that value as the rbh_value of the rbh_value_pair.

Currently, we can only retrieve the stripe count of that directory
(using the key "stripe count"), but the stripe size and pattern will be
added later.

Change-Id: I386e8dd7a4af2c4708b92bc7ebb7358edb31530f
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>