lustre: retrieve stripe size in lustre_get_attribute
This patch adds the stripe size of the attributes that can be retrieved
by the Lustre backend from external tools. It is called by
rbh-find-lustre to get the stripe size of the mount point and compare it
to the requested value for filtering.

Also update the test meson of rbh-find-lustre since the stripe size test
should pass now.

Change-Id: Ibadd9f274dc7215e1a8c0a2183dcb8c43d1892d9
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>