mpi-file : add the mpi-file backend
This patch adds the new mpi-file backend. This backend will be used to
create a mpifileutils file with all the information recover while
syncing another backend (update).

Moreover, this backend will be able to to synchronize a mpifileutils
file with another backend (filter).

All operations described will be added in future patches.

Finally, to use it you have to give the path where is the mpifileutils
file or where to create it in the RBH URI: rbh:mpi-file:path.

Change-Id: I49e30cc208df4ad77d6a0c023f8c8d3846dc5895
Signed-off-by: Gauthier EVRAERD <gauthier.evraerd.ocre@cea.fr>