mpi-file: add update operation
This patch adds the update operation to mpi-file backend. To do the
update, we create an empty flist which we fill with the fsevents
generated by rbh-sync.

The RBH_ID of the last fsentry is stored to know if we need
to create a new entry in the flist or update the last one.

Also, when rbh-sync has finished to sync and there is no data
(ENODATA), we call rbh_backend_update another time to tell to the
backend to save the flist in a file. The flist is saved in binary
format.

Change-Id: I55ed96f91be30f891e15cb876676e3007373a176
Signed-off-by: Gauthier EVRAERD <gauthier.evraerd.ocre@cea.fr>
Reviewed-on: https://gerrithub.io/c/robinhood-suite/robinhood4/+/1194076
Tested-by: VALERI Yoann <yoann.valeri@cea.fr>
Reviewed-by: VALERI Yoann <yoann.valeri@cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>