lustre: fix invalid component count for stripped directories
This patch fixes an issue with stripped directories. The issue is that
those directories have multiple components, but the way to retrieve that
number only worked for regular files.

Therefore, Lustre would tell us the entry has a composite layout, but we
couldn't get the number of components, leading to an invalid allocation
of the number of components, and an invalid retrieval of the striping
values.

To fix this issue, we simply call the Lustre function to iterate over
each components preemptively, and count the number of times we go
through that function, then do the normal allocation/loop.

Change-Id: I086e3300c6ab9f72a822f2e75b5312860530f8a4
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>