lustre: add pattern of available attributes
This patch adds the pattern in the attributes that can be retrieved by
the Lustre backend from external tools. It is called by rbh-find-lustre
to get the pattern of the mount point and compare it to the requested
value for filtering.

Also update the test meson of rbh-find-lustre since the pattern test
should pass now.

Change-Id: Iee33aecda442c9aab74f49ce24852ce4ed4191d8
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>