mpi-file: add filter operation
This patch adds the filter operation to the mpi-file backend. The
filter operation is used when the backend is used as a source. The
fsentry are build with the information we have in the mpifileutils's
file. In the rbh_statx structure, we retrieve the atime, mtime and
ctime, the size, uid, gid and mode.

Also, the ID of each fsentry is build with the relative path of each
entry in the file.

Finally, this patch adds to the mpi_iterator structure two new elements.
The first is a callback use to build a new fsentry by
mpi_iterator_next(). The callback is setup at the creation of the
iterator. The second element is a boolean to indicate to the iterator
if we want to build the rbh_id with the file descriptor or the path of
an entry.

Change-Id: I86df9c0f52bc12fd3695f4444d6fce092ad1a892
Signed-off-by: Gauthier EVRAERD <gauthier.evraerd.ocre@cea.fr>