rbh-sync: add tests for mpi-file backend filter
This patch adds the tests for the mpi-file backend to be used as a
source. Synchronization is done with a file from dwalk and a file
built from the mpi-file backend with rbh-sync. We test if we fill in
the information in mongo correctly. We test the size, path, atime,
mtime and ctime.

Change-Id: I1412a39359ede840988f79a5908691a2d7c5dff2
Signed-off-by: Gauthier EVRAERD <gauthier.evraerd.ocre@cea.fr>
Reviewed-on: https://gerrithub.io/c/robinhood-suite/robinhood4/+/1194773
Reviewed-by: VALERI Yoann <yoann.valeri@cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Tested-by: VALERI Yoann <yoann.valeri@cea.fr>