fsevents: add `--user` option
This patch adds the `--user` option that will be used to acknowledge the
changelogs for the Lustre backend. This is necessary because the ack
function of Lustre requires a user, but we have no way to get it from
the API, so it must be given by the caller.

Incidentally, if a user is not given, the acknowledgement will not be
possible, which we will keep as a valid usage of the command.

Change-Id: Ia0f32db1613c041860afa0b8be0a692deeab8055
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>