fsevents: remove flush size option
This patch removes the flush size options of rbh-fsevents because with
the acknowledgement, flushing only part of the deduplication pool would
mean we wouldn't be able to ack the more regularly accessed entries.
This would make the Lustre changelogs stay on the MDT for too long, and
their management would start lagging behind.

Removing this option thus means that when we flush the deduplication
pool, we will flush all of its entries, so we are able to simply ack all
the changelogs up to the last one handled by the source.

Change-Id: Iec0aa7d4dc342460c25acb40ecc8a9d22d3ba821
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>
Reviewed-on: https://review.gerrithub.io/c/robinhood-suite/robinhood4/+/1194738
Reviewed-by: Gauthier Evraerd <gauthier.evraerd.ocre@cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>