config: implement option searching
This patch implements the different functions to search a key in the
configuration file. It will give back a YAML event corresponding to the
value of the requested key. This event will then have to be parsed by
the caller because only it knows what type it should be.

These functions handle searching in submaps, with a given key containing
`/` between each key. In this case, it will send back the YAML event
associated with the final key.

Change-Id: I4f6a31d24d3542f4ff39f280bf337c4015e978f9
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>