lfind: add `pool` predicate
This patch creates the `pool` predicate which will filter entries based
on a given pool. Internally, it uses the `shell_regex2filter` function
of rbh-find to do the actual filter creation.

For this purpose, `shell_regex2filter` prototype's is changed from
having an `enum predicate` to a `struct rbh_filter_field` to allow
out-of-rbh-find use, while the call in rbh-find is changed to
`regex2filter` which calls `shell_regex2filter` in turn.

Change-Id: Ib338b759358fd11f8f006df35fdb1367238225d8
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>
Reviewed-on: https://review.gerrithub.io/c/robinhood-suite/robinhood4/+/1195405
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>