backends: remove 'skip_error' from rbh_backend_update
This patch removes the 'skip_error' parameter from the
`rbh_backend_update` function. The reason is that the source backend for
rbh-sync and the enrichers of fsevents are already doing that check, so
1) it is a duplicate check, 2) it conflicts with the `--no-skip`
options of rbh-sync and rbh-fsevents, and 3) the source is the one which
knows which fsevent it should return or skip, so if NULL is returned to
the updater, it is either the end of the fsevents available or an error.

Similarly, it removes this same verification from the file sink of
rbh-fsevents.

Change-Id: I9beb08bb9b41017f53e05002aadab00c40862909
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>