dedup: fix invalid copy of empty values
This patch fixes an issue with the copy of rbh_values in the
deduplicator. Previously, the deduplication would assume each value
copied is not NULL, so it would copy its content to the destination.
However that assumption is incorrect, as the YAML source or the Lustre
source will generate events with values that are empty, only to be
filled later by the enricher.

So if we dump the fsevents from the source reader without enriching
them, we want to dump empty values, but the deduplicator prevents that
by creating a new value and trying to copy the old content to it.

Change-Id: I766549035e81bea86e02d31d7acd16ec61517371
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>