find: move shell2pcre to librobinhood
This patch moves the function shell2pcre from rbh-find to
librobinhood. Initially, rbh-find was converting the shell pattern
from the name/path predicate into a Perl Compatible Regex (PCRE) for
the Mongo backend, but the mpifile backend didn't support it. Now,
it is no longer rbh-find but the backend itself that handles and
converts it if needed.

Change-Id: I093bf5469cec829f09c4398aafd5eaa9ecc614f6
Signed-off-by: Gauthier EVRAERD <gauthier.evraerd.ocre@cea.fr>