mpi-file: remove absolute path with -path predicate
This patch removes the absolute path of an entry to use the
relative path like all the other backends. The function MFU_PRED_PATH
from mpifileutils has been redefined. The new function removes the
root from the names stored in the flist before testing the predicate.
The length of the root is retrieved during the filter operation.

Change-Id: I4feb55f1e05efa46e930053027ef04f9990174cb
Signed-off-by: Gauthier EVRAERD <gauthier.evraerd.ocre@cea.fr>