mpi-file: modify time filter
This patch modifies the [acm]time and [acm]min filters with the mpi-file
backend. The time specified in seconds in the filter is now converted to
minutes relative to the current time to use mpifileutils's comparison
functions (MFU_PRED_AMIN, ...) because mpifileutils converts the time
value into the filter unit.

Also, the time is always reduced to minutes even with the [acm]time
filter because it's the smallest time unit that mpifileutils supports.
This avoids sending information about whether it's a [acm]time or
[acm]min filter to the backend.

Change-Id: I7865cf8275219cf60d12bf730101922ef11bc574
Signed-off-by: Gauthier EVRAERD <gauthier.evraerd.ocre@cea.fr>
Reviewed-on: https://gerrithub.io/c/robinhood-suite/robinhood4/+/1196206
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Tested-by: VALERI Yoann <yoann.valeri@cea.fr>
Reviewed-by: VALERI Yoann <yoann.valeri@cea.fr>