config: search in config and environment variables
This patch changes the behaviour of `rbh_config_find` to also check the
environment variables for the requested key if it isn't found in the
configuration file.

Also, it adds a parameter to this call, in order to check the requested
key has a value of a certain type.

However, this type check is only done if the key is found in the
configuration file. For the environment variable, the expected type can
only be string.

Change-Id: I78631bfb31490feb067468c86b9df9e0b1af6fa3
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>
Reviewed-on: https://review.gerrithub.io/c/robinhood-suite/robinhood4/+/1196218
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Gauthier Evraerd <gauthier.evraerd.ocre@cea.fr>