mpi-file: modify creation of mfu_pred
This patchs modifies the way the mfu_pred are created from a
rbh_filter in order to support the OR and NOT logical operators.

The structure of the rbh_filter is reproduced with mfu_pred structure.
Like the rbh_filter, there are logical mfu_pred and comparison
mfu_pred. Each mfu_pred takes a function and an argument. The function
determine if it's a logical or comparison mfu_pred.

The logical mfu_pred are combinations of other mfu_pred.
The functions used for the logic are _MFU_PRED_AND and _MFU_PRED_NULL.
It takes as an argument another mfu_pred containing a combination of
other mfu_pred.

The comparison mfu_pred represents a single predicate.
The functions used are _MFU_PRED_SIZE, _MFU_PRED_PATH, and all the
functions from mpiFileUtils. They takes as an argument the predicate's
value.

Change-Id: I4985ef9b59dfb82b1ade8282b0068c4356738eea
Signed-off-by: Gauthier EVRAERD <gauthier.evraerd.ocre@cea.fr>