mpi-file: add the logical operator OR
This patch adds the logical operator OR to the mpi-file backend. The
logical function _MFU_PRED_OR is added. This function works similarly
to the mfu_pred_execute function but with the difference that as soon
as a predicate returns the value 1, we stop going through the list of
predicates.

Change-Id: I3198728ae6786b80f9375f077ff366832ff66e12
Signed-off-by: Gauthier EVRAERD <gauthier.evraerd.ocre@cea.fr>