tests: check retention stays correct after sync
This patch adds a test to the retention test script that checks that the
retention date stays valid even after a sync.

The issue is that if we sync a directory with a retention attribute,
update the modify time of a file in that directory, then run the retention
script, the directory's expiration date will be updated. However, if we
run a new sync after, the directory's expiration date is set back to the
original date, overwriting any change made by the retention script.

Change-Id: Ibb67acc2ce495fb117f07ba8069097a1f4d1e443
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>