Capabilities: Add a new command to show capabilities of backend.
Add a new command rbh-capabilities to show capabilities of backend among filter,
update and branch. To do this we add one function that load a backend to recover
capabilities and one other that translate binary capabilities indicator to human
readable text. We modify all backend structure to add the capabilities binary
indicator.

Change-Id: Ib14650e5f6d77800d8df286f03b654cd5816bb7e
Signed-off-by: Arnaud GAUTHERET <arnaud.gautheret@cea.fr>
Reviewed-on: https://review.gerrithub.io/c/robinhood-suite/robinhood4/+/1197171
Tested-by: VALERI Yoann <yoann.valeri@cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: VALERI Yoann <yoann.valeri@cea.fr>