retention: set the expiration date from the script
This patch adds the set of the expiration date as a xattr by the
retention script when it is updated. As such, we modify the Lustre
backend to check whether this attribute is already set or not, and if it
is, we do not create it again, but simply change its typing to stay
consistent with other tools' usage of this attribute in the database.

Also do a bit of cleaning of the main retention function of the Lustre
backend.

Change-Id: Ic46a9c875bd9d7fb3460777101a4b33c894c5e1d
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>