retention: compare dir's calculated and set retention
This patch changes the behaviour of the retention if a directory already
has a set expiration date. If that is the case, we calculate the
expiration date using the directory's retention date and it's modify
time, and compare it to the set expiration date. If the result is
greater, we change the expiration date of the directory to the
calculated one, otherwise we keep the already set one.

Change-Id: I52d3a53f1083be8560a650bcafe2670a68f49736
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>