posix: check xattr typing from map
This patch adds the retrieval of a xattr's typing if present in the
xattrs typing map set by the config. If it isn't in the map, or there is
no map given in the config (or no config at all), the xattr's typing
defaults to binary.

Currently, the typings handled is only binary, but others will be added
in the next patches.

Change-Id: I11ff3d734d81a56d2ceb71db6ca801a161698c36
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>
Reviewed-on: https://review.gerrithub.io/c/robinhood-suite/robinhood4/+/1197995
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>