packaging: release patch for RobinHood4 version 1.0
This patch updates the different meson build files and dependencies for
the release of the RobinHood4 toolsuite, version 1.0.

It is not a pure replacement of RobinHood3 yet because there are still
differences in functionnalities, but since there are conflicting tools,
we enforce the fact that both versions cannot be installed on the same
node.

Change-Id: Id56f7abf8bc62fe661108d93eda6e9aa0d0642dd
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>