tests: add suite name to meson tests
Now that the meson configs are unified, all the tests are run by
`meson test`. This patch adds a `suite` parameter to all tests defined
as the name of the sub project. This allows running all the tests of one
tool.

For example, to run all the tests of librobinhood:

meson test --suite librobinhood

Change-Id: Idc896cf316fef44a953109d0581363412e43512d
Signed-off-by: Courrier Guillaume <guillaume.courrier@cea.fr>
Reviewed-on: https://review.gerrithub.io/c/robinhood-suite/robinhood4/+/1199660
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>