bugfix: fix non-regression issue in rbh-sync
This patch fixes an issue with the non-regression of rbh-sync. The issue
was that for the sync posix/mongo, we would test the skipping feature
with a secondary user created in the test, and having it sync entries
with the main user's library path.

However, with the changes to the meson build, the library path was
missing the librobinhood library location, and it didn't have proper
right access to the built files.  Therefore, that user couldn't
synchronize entries, and the test would fail.

Change-Id: I53321091a0938a39df3f8bad193280c9083f286e
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>
Reviewed-on: https://review.gerrithub.io/c/robinhood-suite/robinhood4/+/1201964
Reviewed-by: Gauthier Evraerd <Gauthier.EVRAERD@cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>