find: create function to gather size parameters
This patch adds a function to gather the size parameters when dealing
with a size predicate. Those parameters include the operator (a plus or
minus sign, or nothing), the unit size (G(iga), T(era), w(ord), ...) and
the size itself (any number of digits).

It is mostly a patch to prepare for the next one which will include the
usage of the `$elemMatch` operator of Mongo for filtering on arrays.

Change-Id: Iaceba6d652c6a8cf5f4dee132cc7de821860b2a2
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>