lfind: allow interval filtering in `comp-start` predicate
This patch adds a way for the `comp-start` filtering to do filtering on
an interval.

Since `-comp-start +10M -comp-start -20M` are treated separately, they
will not match all entries that have a component start both superior to
10M and inferior to 20M, but rather entries that have a component start
superior to 10M and another inferior to 20M, with no guaranty of them
being the same.

To circumvent this issue, we now allow the predicate to take in a CSV
string composed of 2 parts, both corresponding to interval bounds.
Therefore, to do the aforementionned filtering, we can now do
`comp-start 10M,20M` and it will work with the interval `[9M+1; 20]`.

Change-Id: I4536a490269c292276ac20f89c9b3beebffc3f3a
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>