bugfix: fix invalid check on retention attribute
This patch fixes an invalid check that was done after retrieving the
retention attribute from the file's xattr. The problem was that we would
convert the expiration date set by RBH to an integer using strtoul, and
then we would check the `str_end` pointer was pointing to a NULL byte.

However, since that is not guaranteed, we would sometime fail the check.
To fix this issue, we instead check that the difference between the
start pointer (given to strtoul) and the end pointer (returned by it) is
not different from the size of attribute we know.

If it is, that means we converted more/less bytes than expected to an
integer, and there is an issue with the attribute. If the result is
equal, we converted the full attribute, nothing more, nothing less, and
we can continue normally.

Change-Id: Ie59b28f164174a8c62a4c6689a8180782c140723
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>