capabilities: bugfix listing not loadable backend
We are removing the fetch backends action in
usr/local/lib64 to prevent rbh-capabilities -l
from showing backends that cannot be loaded.
We remove the fetch action from backends in
usr/local/* to be consistent with how dlopen
works and thus prevent rbh-capabilities -l
from displaying backends that cannot be loaded.
(besause not in the scope of dlopen)

Change-Id: I9dc5c318edc4d2452b835d40a6b1dfb8f3039fe5
Signed-off-by: Arnaud GAUTHERET <arnaud.gautheret@cea.fr>