report: add group stage for agreggation
This patch adds a parameter to the bson pipeline creation function to
set the group stage or not. It is currently only true if the function is
called from the report backend ops. Add the corresponding file and
function.

This additional parameter will be changed in a later patch for a generic
one.

Change-Id: Ia5d99f6d275715582b32af23cdfc435bea70018a
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>