report: add basic group stage to the bson pipeline
This patch adds a basic group stage to the bson pipeline. The goal is
mostly to have a first version that works, then makes everything generic
for additional parameters and uses.

For now, this basic group stage will simply group all the entries in the
backend and sum their sizes.

Change-Id: I18cd5219f0266b167cb1384088322de4cf2b5849
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>
Reviewed-on: https://gerrithub.io/c/robinhood-suite/robinhood4/+/1203362
Reviewed-by: Gauthier Evraerd <Gauthier.EVRAERD@cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>