report: add basic projection stage to the bson pipeline
This patch adds a basic projection stage to the bson pipeline in case we
come from the report function. The goal is mostly to have a first
version that works, then makes everything generic for additional
parameters and uses.

For now, this basic projection stage will simply remove the id from the
output and show the result of the sum done in the grouping stage.

Change-Id: Ic2086cb271c6def0126c8b4350c48b14bbc652f8
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>
Reviewed-on: https://gerrithub.io/c/robinhood-suite/robinhood4/+/1203363
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Gauthier Evraerd <Gauthier.EVRAERD@cea.fr>