mongo: append range key
This patch adds the set of the range key in the bson document. This key
will be the name of the field added to the bson pipeline and the one the
group id will use.

To append this key, this patch adds a function to create a generic key
from the field we want to use. It will be reused by the group stage.

Change-Id: Iaadfcc7e2841b53ae93551a8444a359ec6cf614d
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>
Reviewed-on: https://review.gerrithub.io/c/robinhood-suite/robinhood4/+/1205331
Reviewed-by: Gauthier Evraerd <Gauthier.EVRAERD@cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>