report: search for content map
This patch changes the behaviour of the report output conversion:
instead of skipping every entry in the bson straight to the content map,
we iterate through the bson output, and check the key obtained.

Currently, we only search for the content map, and convert it as done
before. Next patches will add the retrieval of the id map.

Change-Id: I37601348067739f61329879730cd449c4e6e6ac6
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>
Reviewed-on: https://review.gerrithub.io/c/robinhood-suite/robinhood4/+/1205464
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Gauthier Evraerd <Gauthier.EVRAERD@cea.fr>