mongo: add function to group two maps in another one
This patch adds a function to group two maps in another one. It will be
used to group the id map and content map into a final one sent back to
the rbh-report tool when dealing with rbh-report requests.

Currently unused, it will be when we starting retrieving the id in each
document outputted by the rbh-report request.

Change-Id: I175f6e04264c1a2b32e62d6bd28a0afd03ee1e87
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>