mongo: convert report output as value
This patch changes the conversion of the rbh-report output from a map to
a value. Since we want to add the `ids` of each field retrieved from
Mongo, we're not assured they correspond to a map.

Currently, we still only retrieve the content map, so we still assume
the value is a map, but that will be changed.

Change-Id: I2233e3bee41bf20c9d0e410fe3dc687ad910ad05
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>