mongo: retrieve the id from the report results
This patch adds the retrieval of the id in the report resulting
documents. They will be outputted by the report tool in order to know
what the aggregated information correspond to.

Two tests have to be modified to handle the new output format, they will
be in the next patch.

Change-Id: I1e4fffc8361408fc4e02a077a38aa0145813b4a5
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>