report: add result columns structure and init
This patch adds a `result_columns` structure and its init functions. The
goal of this structure will be to hold the input group-by and output
requested fields to be displayed when print results from a rbh-report
command, and the length of the associated columns. With this, we can
build a proper output that is more human readable than a simple CSV.

Change-Id: I45272c8146f08cb0eb8c1002b1e40bcfe3033daf
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>
Reviewed-on: https://review.gerrithub.io/c/robinhood-suite/robinhood4/+/1205870
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Gauthier Evraerd <Gauthier.EVRAERD@cea.fr>