mongo: change default range value for sorting
This patch changes the default value when trying to use ranges in
rbh-report. The issue is that if we use a string as default value, the
sorting of entries will put that string as the first output. That's
because for Mongo, strings have a higher priority than arrays, and thus
are sorted before.

Change-Id: I55f5abcad45063dc99d82e0d42a4738a9fa4b64d
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>