report: change dump functions to write to buffer
This patch changes the `dump_value` function to instead write to a
buffer. This will allow either printing the value or getting its size,
which will be necessary for pretty printing.

Moreover, we separate the `dump_value` function from the other ones, as
the latter are more CSV oriented, while the former is more generic and
will be used by other types of printer.

Change-Id: Id52c794c09f0e0ea6f991ddc04cb1022710deed0
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>