adr: add support for plugin extensions
The goal of this ADR is to suggest the support of plugin extensions.
The idea is that functionnalities such as lustre or retention enrichment
or custom iterators such as MPIFileUtils can be implemented as extensions
to the POSIX backend. Then a backend can use the configuration to
decide which extensions to load and use. These extensions will be
implemented as libraries that can be dynamically loaded by librobinhood.

Extensions will be specific to the backend that will load them. The
first use case will be to refactor the lustre, mpi based and posix
backends to reduce code duplication and clearly identify dependencies
between these backends.

This will also allow backends such as lustre or mpi to be compiled
without the sources of the posix backend. The posix backend will have
to expose some of its internals in a header file. This header file can
then be installed as part of a robinhood4-posix-devel RPM for instance.
This RPM alongside robinhood4-devel will be the only dependencies
required to build external extensions.

Once this is implemented, librobinhood will have the backends posix,
mpi-file, mongo and hestia, and the extensions lustre, mfu, retention.
The lustre, mfu and retention backends will simply be extensions to the
posix backend. The posix-mpi and lustre-mpi backends won't be necessary
anymore.

Change-Id: I2b9c9c581a8d4233edbd0735f3237cf0765b15f3
Signed-off-by: Guillaume Courrier <guillaume.courrier@cea.fr>