backend: lookup config to find custom backends
backend_new will now look at the config and search for values under the
backends key. This will allow custom backends to be defined. The
configuration looks like this:

---
backends:
    test:
        extends: "posix"
---

If we give the URI "rbh:test:/mnt", backend_new will look into the
configuration and look for the string associated to "backends/test/extends".
If it exists, it will load the specified plugin. If not, it will try to
load the "test" plugin. Which in this case does not exist. But for the
posix backend for instance, this would have the same behavior as
previously. This also leaves the possibility to redefine existing
plugins. For example, this would be allowed:

---
backends:
    posix:
        extends: "posix"
---

This can be useful if we want to extend the defaults of an existing
plugin but keep its name for simplicity.

Change-Id: I582a3afd1c8d6ae73382102c61ef867ef165f059
Signed-off-by: Guillaume Courrier <guillaume.courrier@cea.fr>
Reviewed-on: https://review.gerrithub.io/c/robinhood-suite/robinhood4/+/1210946
Reviewed-by: VALERI Yoann <yoann.valeri@cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Gauthier EVRAERD <Gauthier.EVRAERD@cea.fr>