backend: add type as an argument to rbh_backend_plugin_new
This new parameter will contain the name of the backend that was given
in the URI. This will allow backends to fetch elements from the
"backends" key in the configuration to know how they should behave.

For instance, with this configuration:

---
backends:
    test:
        extends: posix
        enrichers:
            - retention
---

The posix backend will be created with type = "test", then get the value
associated to "backends/test" to load the retention plugin.

Change-Id: I9a5573937433feaa1dc05b7bef9208ace6eca3ff
Signed-off-by: Guillaume Courrier <guillaume.courrier@cea.fr>
Reviewed-on: https://review.gerrithub.io/c/robinhood-suite/robinhood4/+/1210947
Reviewed-by: Gauthier EVRAERD <Gauthier.EVRAERD@cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: VALERI Yoann <yoann.valeri@cea.fr>