posix: load enrichers when building the backend
This patch adds a list of enrichers in the structure posix_backend.
This structure is initialized by looking for backends/<name>/iterator
and backends/<name>/enrichers in the configuration by the function
rbh_posix_backend_new.

Since custom iterators are not yet implemented, this patch only loads
the iterator if necessary but does nothing with the loaded extension.

Change-Id: Ia87d355d756c8c7968d69af3472d4eb886891c1d
Signed-off-by: Guillaume Courrier <guillaume.courrier@cea.fr>
Reviewed-on: https://review.gerrithub.io/c/robinhood-suite/robinhood4/+/1210949
Reviewed-by: Gauthier EVRAERD <Gauthier.EVRAERD@cea.fr>
Reviewed-by: VALERI Yoann <yoann.valeri@cea.fr>