lustre: implement lustre enricher
This patch implements the new enricher API for the Lustre backend. The
POSIX backend can now load the lustre enricher and call its enricher
to get lustre specific attributes. This enricher has now the role of
the rbh_backend_get_attribute method as well as the old
inode_xattrs_callback. The next step will be to implement the
get_attribute method for the POSIX backend so that it can call its
enrichers. Once this is done, the lustre backend won't be useful
anymore.

For now, the Lustre enricher is still a fully functionning backend as
well as both a Lustre and POSIX extention. It is still a Lustre
extention to keep the compatibilty with the current implementation of
the Lustre backend and pass the tests.

Change-Id: Ic6036130a95f277dcf569f302c7d69528cd1fcd4
Signed-off-by: Guillaume Courrier <guillaume.courrier@cea.fr>
Reviewed-on: https://review.gerrithub.io/c/robinhood-suite/robinhood4/+/1210951
Reviewed-by: Gauthier EVRAERD <Gauthier.EVRAERD@cea.fr>
Reviewed-by: VALERI Yoann <yoann.valeri@cea.fr>