rbh-fsevents: add a list of extension enrichers to the posix enricher
To prepare for the removal of the retention enrichment from the lustre enrichment,
this patch introduces a list of enricher functions to the posix enricher.
These functions will be called on each xattr to enrich so that the individual
enrichers can decide if they need to fetch data or not.

Change-Id: I13132f934d0b8c4720ed0f672e2bcbfd1c66505b
Signed-off-by: Guillaume Courrier <guillaume.courrier@cea.fr>