filter: add GET operation/filter
This patch adds the new GET operation and a new type of rbh_filter,
the get filter. The get filter can be used when we need to build a
filter with a value that requires to be fetch from a fsentry of a
backend.

This filter contains two comparison filters. One for
the fsentry to retrieve in the backend and one for the real filter.

Change-Id: Iff388d1599bdc5064c2115f527275b1e5eef2d52
Signed-off-by: Gauthier EVRAERD <gauthier.evraerd@cea.fr>