find: add a prefetch to complete filters
This patch adds a check that verify if a filter need to be completed
before doing the find. If a filter need to be completed, we iterate all
the filter to find the filter to complete. To complete the filter, we
do a prefetch to retrieve the fsentry and complete the filter.

It can slow down the find but it's insignificant compared to the real
find.

Change-Id: Ia4b26904392c44c9d1e1eb1f3d9640b6e6a21be7
Signed-off-by: Gauthier EVRAERD <Gauthier.EVRAERD@cea.fr>