find: add `-regex` and `-iregex` filters
These filters allow one to filter entries based on a regex. The
filter tries to match the regex to the whole path of entries. The
-iregex filter is like the -regex filter, but the filter is case
insensitive.

Also, the options RBH_RO_SHELL_PATTERN have been added to the regex
options to differentiate filters that take as input a shell pattern or
a regex pattern.

Change-Id: I3e08ac3bb3c7f3d6f3621567c8d99c79c3c86db0
Signed-off-by: Gauthier EVRAERD <Gauthier.EVRAERD@cea.fr>
Reviewed-on: https://gerrithub.io/c/robinhood-suite/robinhood4/+/1208975
Reviewed-by: VALERI Yoann <yoann.valeri@cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>