rbh-fsevents: implementation of generic enrichers
To make enrichers generic, we need to let them see
all the events that we manage and what the source
wants to update. This patch adds a struct enrich_request
that represents what the source requested to be enriched.

Currently, only two types are necessary: statx and
xattr. FSEvents enrichers will be called for each
events of either type. They can let the caller know
if they don't handle this event by returning -1
with errno = ENOTSUP. For statx events, the statx
mask is passed to the enricher to let it know which
meta data has changed on the entry according to the
source. For xattr events, the name of the xattr is
given.

The lustre enricher only supports the "lustre" and
"path" xattr. It does nothing on statx events.

The retention enricher supports both statx and
xattr. For statx, it will update the retention if
the atime, ctime or mtime has changed. For
xattr events, only the retention attribute as
defined in the configuration will trigger an
update.

This patch also fixes the lack of support for the
skip option in the posix enricher. The tests were
previously only running with the lustre enricher.
But now that both the posix and lustre enrichers
share the posix code, this test was failing.

If extra enrichers are required in the future,
only the custom fsevent enricher callback will
need to be implemented. The posix enricher will
simply look into the configuration to know which
enrichers to use.

Enrichers are still "hardcoded" into rbh-fsevents
as they are not loaded dynamically.

Change-Id: Ie16d981dbc23f0b30c2a049991a4f138d69d2f84
Signed-off-by: Guillaume Courrier <guillaume.courrier@cea.fr>
Reviewed-on: https://review.gerrithub.io/c/robinhood-suite/robinhood4/+/1210968
Reviewed-by: VALERI Yoann <yoann.valeri@cea.fr>
Tested-by: VALERI Yoann <yoann.valeri@cea.fr>
Reviewed-by: Gauthier EVRAERD <Gauthier.EVRAERD@cea.fr>